+++ src/webgl/lineFeature.js
@@ -461,15 +461,12 @@ var webgl_lineFeature = function (arg) {
   };
 
   /**
-   * Return list of vgl actors used for rendering.
+   * List vgl actors.
    *
-   * @returns {vgl.actor[]}
+   * @returns {vgl.actor[]} The list of actors.
    */
   this.actors = function () {
-    if (!m_actor) {
-      return [];
-    }
-    return [m_actor];
+    return m_actor ? [m_actor] : [];
   };
 
   /**