+++ tests/gl-cases/svgLines.js
@@ -47,5 +47,5 @@ describe('svgLines', function () {
 
       imageTest.imageTest('svgLines', null, 0.0015, done, myMap.onIdle, 0, 2);
     });
-  });
+  }, 30000);
 });