+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changes
 - The point clustering radius value is now in display pixels (#983)
 
+### Bug Fixes
+- Fixed drawing partial fixed-scale canvas quads (#985)
+
 ## Version 0.19.2
 
 ### Improvements