+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Layers that use webgl renderers automatically share contexts when possible.  Layers can switch renderers manually as well.  This largely avoids the limitation of number of webgl contexts in a browser.
 - Support affine transforms in the proj4 string (#986)
 
+### Improvements
+- Spped up rendering geojson features by using constant values for constant geojson styles (#987)
+
 ### Changes
 - The point clustering radius value is now in display pixels (#983)